rename network -> p2p_network
