+++ src/p2p_network/connection.py
